C39 - Bug Fix
Fixed a validation bug on the frontend, where now only a "menedzer" (admin) can look at the list of employees and interact with it. Fixed another bug where when you update a user with an image from the admin panel, the image will not be deleted, but it will stay the same.