C73 - Checkbox, Review tab, Profile picture BugFix
Fixed the "bugs" regarding having 2 buttons for choosing is a dish a specialty or not. Now those 2 buttons are simply 1 checkbox. The review tab now has the option to choose a dish all the time without the need to go back to the menu tab. Also now you can upload a profile picture too when registering or editing a user.